requirements.txt
tensorflow-gpu~=2.6.0
numpy~=1.19.4
PyCryptodome~=3.14.1
scipy~=1.8.1
matplotlib~=3.5.2
absl-py~=0.12.0